requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pdfplumber==0.11.0
chromadb==0.5.0
openai>=2.0.0
python-multipart==0.0.9
pydantic==2.7.1
numpy==1.26.4
python-dotenv==1.0.1
